# Documentacion/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Vida Cotidiana API
  version: 1.0.0
  description: >
    Canonical API contract for Vida Cotidiana V1. All timestamps are
    ISO-8601 in UTC. All identifiers are UUIDv4. Authentication is delegated
    to Keycloak (OIDC/OAuth 2.1, Authorization Code + PKCE, DEC-004/ADR-008);
    this API is a pure OAuth2 resource server and defines no /auth/* login
    endpoint of its own. An operational healthcheck (e.g. Spring Boot
    Actuator /actuator/health) exists outside this versioned business
    contract and is intentionally not documented here (see
    07-backend-architecture.md).
servers:
  - url: /api/v1
security:
  - keycloakBearer: []
paths:
  /me:
    get:
      summary: Get current user
      security:
        - keycloakBearer: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Request account deletion (soft delete, 30-day grace period before purge, DEC-015/ADR-012)
      security:
        - keycloakBearer: []
      responses:
        '202':
          description: Deletion requested; account moves to PENDING_DELETION, purged after 30 days
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/modes:
    post:
      summary: Enable a mode (Personal or Laboral) for the current user (ADR-015/FR-016/UC-15)
      description: >
        Only-additive — enables the given mode, there is no way to disable
        one yet (open, non-blocking TBD in ADR-015). Idempotent: enabling an
        already-enabled mode is a no-op that still returns 200. POST rather
        than PATCH /me — this is an action ("activate this mode"), the same
        style already used by POST /reminders/{id}/complete.
      security:
        - keycloakBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableModeRequest'
      responses:
        '200':
          description: Mode enabled; returns the updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /reminders:
    get:
      summary: List own and shared reminders (paginated)
      security:
        - keycloakBearer: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated reminder list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    additionalProperties: true # BE-035, see PageMeta
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Reminder'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a reminder (caller becomes owner)
      security:
        - keycloakBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reminders/{id}:
    get:
      summary: Get a reminder (owner or active collaborator)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reminder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update reminder (owner only). Requires the current version for optimistic locking.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking) — the reminder was modified concurrently; refetch and retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete reminder (owner only). Cascades to INVITATION/REMINDER_SHARE and notifies active collaborators before deletion (DEC-002).
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /reminders/{id}/complete:
    post:
      summary: Toggle the single global completion state (owner or active collaborator, DEC-001). Optionally accepts the expected version for optimistic locking.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteReminderRequest'
      responses:
        '200':
          description: Completion state toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /notes:
    get:
      summary: List own notes (paginated) — funcionalidad Agenda+Notas completa, V2.
      security:
        - keycloakBearer: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated note list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    additionalProperties: true # BE-035, see PageMeta
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Note'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a note (caller becomes owner)
      security:
        - keycloakBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notes/{id}:
    get:
      summary: Get a note (owner only — no sharing/collaborator concept for this entity)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update note (owner only). Requires the current version for optimistic locking.
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking) — the note was modified concurrently; refetch and retry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete note (owner only)
      security:
        - keycloakBearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    keycloakBearer:
      type: oauth2
      description: >
        OAuth2/OIDC access token issued by Keycloak (Authorization Code +
        PKCE from the client). The backend validates the bearer JWT on every
        request as a pure OAuth2 resource server; it never issues its own
        session or token (DEC-004/ADR-008). Token issuer/JWK Set URL are
        environment configuration, not part of this contract.
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.example/realms/vida-cotidiana/protocol/openid-connect/auth
          tokenUrl: https://keycloak.example/realms/vida-cotidiana/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect authentication
  parameters:
    PageParam:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index.
    SizeParam:
      in: query
      name: size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Page size (max 100).
  schemas:
    Error:
      type: object
      description: Uniform error envelope. Never includes stack traces, SQL, internal names or secrets.
      required: [code, message, traceId]
      properties:
        code:
          type: string
          example: REMINDER_NOT_FOUND
        message:
          type: string
          example: The requested reminder was not found.
        traceId:
          type: string
    PageMeta:
      type: object
      # BE-035: additionalProperties: true is required here because PageMeta is combined via allOf
      # with a sibling object (page items) at each paginated endpoint — without it, strict allOf
      # validators (incl. swagger-request-validator/TEST-API-001) reject each branch as a closed
      # schema and fail the sibling's fields as "not allowed", even though the merged instance is
      # exactly what the composed schema intends. Not a loosening of type/required validation.
      additionalProperties: true
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    User:
      type: object
      required: [id, email, personalEnabled, laboralEnabled]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
          nullable: true
        deletionStatus:
          type: string
          enum: [ACTIVE, PENDING_DELETION, DELETED]
        personalEnabled:
          type: boolean
          description: ADR-015/FR-014/FR-016. Only-additive — there is no matching "disable" yet (open, non-blocking TBD in ADR-015).
        laboralEnabled:
          type: boolean
          description: ADR-015/FR-014/FR-016. Only-additive — see personalEnabled.
    EnableModeRequest:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [PERSONAL, LABORAL]
    CreateReminderRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        dueAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp in UTC.
        context:
          type: string
          enum: [PERSONAL, LABORAL]
          description: >
            ADR-015/FR-019. Optional — inferred by the client from the navbar
            of origin, no selector in the create form. Omitted by every
            pre-ADR-015 caller; the server defaults to PERSONAL when absent
            (same reasoning as the REMINDER.context backfill migration).
        iconId:
          type: string
          maxLength: 32
          description: >
            Funcionalidad Agenda completa (V2). Opcional — identificador
            estable de un catálogo que resuelve el frontend
            (web/src/core/ui/pickers/pickerCatalog.ts), nunca una ruta de
            asset. Mismo contrato que Note.iconId.
        stickerId:
          type: string
          maxLength: 32
        personId:
          type: string
          format: uuid
          description: >
            ADR-016/FR-023. Opcional — debe pertenecer al mismo dueño; si no,
            404 PERSON_NOT_FOUND (nunca revela existencia de un recurso ajeno).
        projectId:
          type: string
          format: uuid
          description: ADR-016/FR-023. Opcional — mismo contrato de autorización que personId (404 PROJECT_NOT_FOUND).
        location:
          type: string
          maxLength: 500
          description: ADR-016/FR-024. Opcional — solo relevante para una "reunión" (context=LABORAL con participantes vía reminders/{id}/shares).
    UpdateReminderRequest:
      type: object
      description: Partial update. version is required to detect concurrent modifications (optimistic locking).
      required: [version]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        dueAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp in UTC.
        iconId:
          type: string
          maxLength: 32
          description: >
            NO es parcial como title/description/dueAt — siempre se aplica
            tal cual se envía (incluida su ausencia, que lo limpia). Mismo
            contrato que UpdateNoteRequest.iconId.
        stickerId:
          type: string
          maxLength: 32
        personId:
          type: string
          format: uuid
          description: ADR-016/FR-023. Parcial (null deja el valor sin cambios) — no existe forma de desvincular una vez asignado en V3.
        projectId:
          type: string
          format: uuid
          description: ADR-016/FR-023. Mismo contrato parcial que personId.
        location:
          type: string
          maxLength: 500
          description: ADR-016/FR-024. Mismo contrato parcial que title/description.
        version:
          type: integer
          description: Expected current version of the reminder; a mismatch returns 409.
    CompleteReminderRequest:
      type: object
      description: Optional. If version is provided, the server validates it for optimistic locking; if omitted, the toggle is applied without a concurrency check.
      properties:
        version:
          type: integer
    Reminder:
      type: object
      required: [id, title, status, context, version]
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        dueAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp in UTC.
        status:
          type: string
          enum: [PENDING, COMPLETED]
          description: Single global completion state shared by owner and all active collaborators (DEC-001).
        context:
          type: string
          enum: [PERSONAL, LABORAL]
          description: ADR-015/FR-017/FR-018/FR-019. Always present in responses — see CreateReminderRequest.context for how it's set.
        iconId:
          type: string
          maxLength: 32
        stickerId:
          type: string
          maxLength: 32
        personId:
          type: string
          format: uuid
          description: ADR-016/FR-023. Vínculo opcional a una Persona propia.
        projectId:
          type: string
          format: uuid
          description: ADR-016/FR-023. Vínculo opcional a un Proyecto propio.
        location:
          type: string
          maxLength: 500
          description: ADR-016/FR-024. Solo relevante para una "reunión".
        version:
          type: integer
          description: Optimistic locking version, incremented on every update.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateNoteRequest:
      type: object
      required: [title]
      description: >
        Funcionalidad Agenda+Notas completa (V2). iconId/stickerId son
        opcionales — identificadores estables de un catálogo que resuelve
        el frontend (web/src/core/ui/pickers/pickerCatalog.ts), nunca una
        ruta de asset ni el emoji en sí.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        iconId:
          type: string
          maxLength: 32
        stickerId:
          type: string
          maxLength: 32
        personId:
          type: string
          format: uuid
          description: ADR-016 Fase 3a/FR-029 (candidato V4). Opcional — debe pertenecer al mismo dueño; si no, 404 PERSON_NOT_FOUND.
        projectId:
          type: string
          format: uuid
          description: ADR-016 Fase 3a/FR-029 (candidato V4). Opcional — mismo contrato de autorización que personId (404 PROJECT_NOT_FOUND).
    UpdateNoteRequest:
      type: object
      description: >
        Partial update para title/description (opcionales, null deja el
        valor guardado sin cambios). iconId/stickerId NO son parciales —
        siempre se aplican tal cual se envían (incluida su ausencia, que
        los limpia); el formulario de edición real siempre reenvía su
        selección completa actual. version es obligatorio (bloqueo
        optimista).
      required: [version]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        iconId:
          type: string
          maxLength: 32
        stickerId:
          type: string
          maxLength: 32
        personId:
          type: string
          format: uuid
          description: ADR-016 Fase 3a/FR-029 (candidato V4). Parcial (null deja el valor sin cambios).
        projectId:
          type: string
          format: uuid
          description: ADR-016 Fase 3a/FR-029 (candidato V4). Mismo contrato parcial que personId.
        version:
          type: integer
          description: Expected current version of the note; a mismatch returns 409.
    Note:
      type: object
      required: [id, title, version]
      description: >
        Funcionalidad Agenda+Notas completa (V2). Owner-only entity — no
        sharing/collaborator concept, mismo criterio que Warranty. Sin
        concepto de status/completado (a diferencia de Reminder/Warranty).
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        iconId:
          type: string
          maxLength: 32
        stickerId:
          type: string
          maxLength: 32
        personId:
          type: string
          format: uuid
          description: ADR-016 Fase 3a/FR-029 (candidato V4). Vínculo opcional a una Persona propia.
        projectId:
          type: string
          format: uuid
          description: ADR-016 Fase 3a/FR-029 (candidato V4). Vínculo opcional a un Proyecto propio.
        version:
          type: integer
          description: Optimistic locking version, incremented on every update.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Input validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated (missing or invalid bearer token).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Caller is authenticated but has no permission on this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource does not exist or is not accessible to the caller (never distinguishes the two, to avoid leaking existence, SEC-001).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...